+++ R/load_ffverse.R
@@ -22,6 +22,8 @@ load_ff_playerids <- function() {
 #'
 #' Accesses DynastyProcess.com's repository of the latest FP expert consensus rankings - updated on a weekly basis.
 #'
+#' @param type one of "draft" or "week", for preseason or inseason-weekly rankings
+#'
 #' @return a dataframe of expert consensus rankings
 #'
 #' @examples
@@ -33,8 +35,17 @@ load_ff_playerids <- function() {
 #' @seealso <https://www.fantasypros.com>
 #'
 #' @export
-load_ff_rankings <- function(){
-  out <- rds_from_url("https://github.com/dynastyprocess/data/raw/master/files/db_fpecr_latest.rds")
+load_ff_rankings <- function(type = c("draft","week")){
+
+  type <- rlang::arg_match0(type, c("draft","week"))
+
+  url <- switch(
+    type,
+    draft = "https://github.com/dynastyprocess/data/raw/master/files/db_fpecr_latest.rds",
+    week = "https://github.com/dynastyprocess/data/raw/master/files/fp_latest_weekly.rds"
+  )
+
+  out <- rds_from_url(url)
   class(out) <- c("tbl_df","tbl","data.table","data.frame")
   out
 }